refactor: `types` directory architecture
P010-39
