+++ packages/grease/src/types/SemanticVersionStable.ts
@@ -1,0 +1,9 @@
+/**
+ * @file Type Definitions - SemanticVersionStable
+ * @module grease/types/SemanticVersionStable
+ */
+
+/**
+ * Stable semantic version string schema.
+ */
+export type SemanticVersionStable = `${number}.${number}.${number}`
