+++ packages/grease/src/types/TVODefaults.ts
@@ -1,0 +1,13 @@
+import type { TransformValidationOptions } from 'class-transformer-validator'
+
+/**
+ * @file Type Definitions - TVODefaults
+ * @module grease/types/TVODefaults
+ */
+
+/**
+ * Default options object for [`class-transformer-validator`][1].
+ *
+ * [1]: https://github.com/MichalLytek/class-transformer-validator
+ */
+export type TVODefaults = Readonly<TransformValidationOptions>
