+++ packages/grease/src/types/Updater.ts
@@ -1,0 +1,15 @@
+import type { UpdaterCustom } from './UpdaterCustom'
+import type { UpdaterJSON } from './UpdaterJSON'
+
+/**
+ * @file Type Definitions - Updater
+ * @module grease/types/Updater
+ */
+
+/**
+ * A resolved updater.
+ */
+export type Updater = {
+  filename: string
+  updater?: UpdaterCustom | UpdaterJSON
+}
