+++ packages/grease/src/types/UpdaterResolver.ts
@@ -1,0 +1,16 @@
+import type { Updater } from './Updater'
+import type { UpdaterCustom } from './UpdaterCustom'
+
+/**
+ * @file Type Definitions - UpdaterResolver
+ * @module grease/types/UpdaterResolver
+ */
+
+/**
+ * Object used to resolve updaters.
+ */
+export type UpdaterResolver = {
+  filename?: Updater['filename']
+  type?: 'json' | 'plain-text'
+  updater?: UpdaterCustom | string
+}
